+++ packages/web-components/src/components/image-with-caption/image-with-caption.ts
@@ -177,4 +177,5 @@ class DDSImageWithCaption extends StableSelectorMixin(ModalRenderMixin(FocusMixi
   static styles = styles;
 }
 
+/* @__GENERATE_REACT_CUSTOM_ELEMENT_TYPE__ */
 export default DDSImageWithCaption;