+++ packages/web-components/src/components/leadspace-with-search/leadspace-with-search.ts
@@ -16,6 +16,7 @@ import '../horizontal-rule/horizontal-rule';
 import StableSelectorMixin from '../../globals/mixins/stable-selector';
 import styles from './leadspace-with-search.scss';
 import { ADJACENT_THEMES } from './defs';
+import StickyHeader from '../../internal/vendor/@carbon/ibmdotcom-utilities/utilities/StickyHeader/StickyHeader';
 
 const { prefix } = settings;
 const { stablePrefix: ddsPrefix } = ddsSettings;
@@ -97,6 +98,10 @@ class DDSLeadspaceWithSearch extends StableSelectorMixin(LitElement) {
     });
   }
 
+  protected firstUpdated() {
+    StickyHeader.global.leadspaceWithSearch = this;
+  }
+
   render() {
     return html`
       <div class="${prefix}--content-layout">
@@ -107,8 +112,10 @@ class DDSLeadspaceWithSearch extends StableSelectorMixin(LitElement) {
         </div>
       </div>
       <div class="${this._getSearchClass()}">
-        <slot name="search"></slot>
-        <div class="${prefix}--sticky-header">${this._heading}</div>
+        <div class="${prefix}--search-container-inner">
+          <div class="${prefix}--sticky-header">${this._heading}</div>
+          <slot name="search"></slot>
+        </div>
       </div>
       <slot name="hr"></slot>
       ${this._contents.map(e => {