+++ packages/web-components/src/components/masthead/masthead.ts
@@ -12,6 +12,7 @@ import settings from 'carbon-components/es/globals/js/settings';
 import ddsSettings from '../../internal/vendor/@carbon/ibmdotcom-utilities/utilities/settings/settings';
 import StableSelectorMixin from '../../globals/mixins/stable-selector';
 import styles from './masthead.scss';
+import StickyHeader from '../../internal/vendor/@carbon/ibmdotcom-utilities/utilities/StickyHeader/StickyHeader';
 
 const { prefix } = settings;
 const { stablePrefix: ddsPrefix } = ddsSettings;
@@ -26,6 +27,10 @@ const { stablePrefix: ddsPrefix } = ddsSettings;
  */
 @customElement(`${ddsPrefix}-masthead`)
 class DDSMasthead extends StableSelectorMixin(LitElement) {
+  firstUpdated() {
+    StickyHeader.global.masthead = this;
+  }
+
   render() {
     return html`
       <div class="${prefix}--masthead__l0">