+++ packages/web-components/src/components/quote/quote-source-bottom-copy.ts
@@ -36,4 +36,5 @@ class DDSQuoteSourceBottomCopy extends LitElement {
   static styles = styles;
 }
 
+/* @__GENERATE_REACT_CUSTOM_ELEMENT_TYPE__ */
 export default DDSQuoteSourceBottomCopy;