+++ packages/services-store/CHANGELOG.md
@@ -3,6 +3,64 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.33.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/services-store/compare/@carbon/ibmdotcom-services-store@1.33.0-rc.1...@carbon/ibmdotcom-services-store@1.33.0) (2022-04-25)
+
+**Note:** Version bump only for package @carbon/ibmdotcom-services-store
+
+
+
+
+
+# [1.33.0-rc.1](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/services-store/compare/@carbon/ibmdotcom-services-store@1.33.0-rc.0...@carbon/ibmdotcom-services-store@1.33.0-rc.1) (2022-04-21)
+
+**Note:** Version bump only for package @carbon/ibmdotcom-services-store
+
+
+
+
+
+# [1.33.0-rc.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/services-store/compare/@carbon/ibmdotcom-services-store@1.31.0...@carbon/ibmdotcom-services-store@1.33.0-rc.0) (2022-04-11)
+
+
+### Bug Fixes
+
+* **deps:** update dependency carbon-components to v10.54.0 ([#8408](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/services-store/issues/8408)) ([588981c](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/services-store/commit/588981c))
+* **deps:** update dependency carbon-components to v10.55.1 ([#8483](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/services-store/issues/8483)) ([579f9a4](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/services-store/commit/579f9a4))
+
+
+
+
+
+# [1.32.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/services-store/compare/@carbon/ibmdotcom-services-store@1.32.0-rc.1...@carbon/ibmdotcom-services-store@1.32.0) (2022-03-28)
+
+**Note:** Version bump only for package @carbon/ibmdotcom-services-store
+
+
+
+
+
+# [1.32.0-rc.1](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/services-store/compare/@carbon/ibmdotcom-services-store@1.32.0-rc.0...@carbon/ibmdotcom-services-store@1.32.0-rc.1) (2022-03-17)
+
+
+### Features
+
+* **scrolling-utility:** Add new utility to mimic dotcom-shell functionality ([#8434](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/services-store/issues/8434)) ([d06b45a](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/services-store/commit/d06b45a))
+
+
+
+
+
+# [1.32.0-rc.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/services-store/compare/@carbon/ibmdotcom-services-store@1.31.0...@carbon/ibmdotcom-services-store@1.32.0-rc.0) (2022-03-16)
+
+
+### Bug Fixes
+
+* **deps:** update dependency carbon-components to v10.54.0 ([#8408](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/services-store/issues/8408)) ([588981c](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/services-store/commit/588981c))
+
+
+
+
+
 # [1.31.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/main/packages/services-store/compare/@carbon/ibmdotcom-services-store@1.31.0-rc.3...@carbon/ibmdotcom-services-store@1.31.0) (2022-02-28)
 
 **Note:** Version bump only for package @carbon/ibmdotcom-services-store